### 2024-05-22 #6769 @editorialbot
[PRE REVIEW]: AutoRA: Automated Research Assistant for Closed-Loop Empirical Research
**Submitting author:** @musslick (<a href="http://orcid.org/0000-0002-8896-639X">Sebastian Musslick</a>)
**Repository:** https://github.com/AutoResearch/autora-paper
**Branch with paper.md** (empty if default branch): main
**Version:** v4.0.0
