### 2024-05-22 #6769 @crvernon
:wave: @musslick - thank you for your submission to JOSS.  It seems the `main` branch is missing the actual `autora` package.  Can you confirm please?  Thanks!
