### 2024-05-23 #6769 @jbytecode
@musslick - Thank you for submitting your work to JOSS. I am the handling editor of this submission. As our topic editor @crvernon states, the `main` branch of the repo contains a test file and some other Python files for documenting. It is a better approach to host the source code in main branch and possibly carrying the paper markdown in a separate branch, e.g., joss-paper. The other thing, adding the folders created by editors, e.g., `.vscode` to `.gitignore` file would be beneficial just because they are user-specific.

Could you please address these issues and ping me again?

Thank you in advance.
