### 2024-05-23 #6769 @musslick
@jbytecode @crvernon Thank you for handling this so quickly!

We apologize for the confusion about the location of our code. The problem is that our code is intentionally distributed across multiple Python packages with a shared namespace. We made this design choice for two reasons. For users, we wanted to minimize time and storage required for individual installations of `autora` (a single installation of `autora` may only require a small subset of packages within the namespace). For contributors, we wanted to minimize dependency conflicts that they have to deal with. `autora` lists all of its packages (our total code contribution) in the [pyproject.tml](https://github.com/AutoResearch/autora/blob/main/pyproject.toml).

However, we do have two required dependencies (`autora-core`) and (`autora-synthetic`). `autora-core` contains most of the workflow mechanic for `autora` and can be found here: [https://github.com/AutoResearch/autora-core](https://github.com/AutoResearch/autora-core). If you prefer the review to happen within a single repository, then we could use this one.

Please let us know about your preference. We are happy to move the paper (and review of the code) to `autora-core` which is a required dependency of `autora` and hosts the code for the main workflow mechanics. I could then host the paper in a separate branch named "joss-paper" in the repository for `autora-core`. Alternatively, we can keep the paper in the "main" `autora` package and I supply a list with the repositories of all relevant subpackages. Either way, I created a separate branch "joss-paper" on the `autora` repository for all future work on the paper.

Our apologies again for this complication! We appreciate your guidance on how to proceed.
