### 2024-05-24 #6769 @jbytecode
@musslick - We talked about this situation with our track editor. Additionally, one of our submissions resembled this situation, we can use it as an example to continue. Please take a look the repository here:

https://github.com/PdIPS/CBX

This repository hosts a JOSS submission, e.g., paper and bibtex. It provides a clean explanation that the submission links to other repositories in which the source code is hosted. So I suggest

- Opening a third repository that holds the submission only
- Please make a clean explanation that the repository hosts a JOSS submission and links to the source code repositories in README.md.

After finishing this procedure, please let me know the new repository so I can change the submission address in the pre-review issue. In your situation, we will have a repository for submission, a repository for `autora-core`, and a repository for `autora-synthetic`, respectively, with last two linked to the first one. We will ask our reviewers to examine and review both of them.

Please ping me if there is anything unclear or if you get into any trouble.

Thank you in advance.
