+++ README.md
@@ -1,0 +1,2 @@
+# autora-paper
+Hosts the JOSS submission for the autora framework.
