+++ paper.md
@@ -1,0 +1,95 @@
+---
+title: 'AutoRA: Automated Research Assistant for Closed-Loop Empirical Research'
+tags:
+  - Python
+  - automated scientific discovery
+  - symbolic regression
+  - active learning
+  - closed-loop behavioral science
+authors:
+  - name: Sebastian Musslick
+    orcid: 0000-0002-8896-639X
+    affiliation: "1, 2"
+    corresponding: true
+  - name: Benjamin Andrew
+    affiliation: 1
+  - name: Chad C. Williams
+    orcid: 0000-0003-2016-5614
+    affiliation: 1
+  - name: Joshua T. S. Hewson
+    affiliation: 1
+  - name: Star Li
+    orcid: 0009-0009-7849-1923
+    affiliation: 3
+  - name: Ioana Marinescu
+    affiliation: 4
+  - name: Marina Dubova
+    orcid: 0000-0001-5264-0489
+    affiliation: 5
+  - name: George T. Dang
+    affiliation: 1
+  - name: Younes Strittmatter
+    orcid: 0000-0002-3414-2838
+    equal-contrib: true
+    affiliation: 1
+  - name: John G. Holland
+    orcid: 0000-0001-6845-8657
+    equal-contrib: true
+    affiliation: 1
+affiliations:
+  - name: Brown University, USA
+    index: 1
+  - name: Osnabrück University, Germany
+    index: 2
+  - name: University of Chicago, USA
+    index: 3
+  - name: Princeton University, USA
+    index: 4
+  - name: University of Indiana, USA
+    index: 5
+date: 22 May 2024
+bibliography: paper.bib
+
+---
+
+# Summary
+
+Automated Research Assistant (`autora`) is a Python package for automating and integrating empirical research processes, such as experimental design, data collection, and model discovery. With this package, users can define an empirical research problem and specify the methods they want to employ for solving it. `autora` is designed as a declarative language in that it provides a vocabulary and set of abstractions to describe and execute scientific processes and to integrate them into a closed-loop system for scientific discovery. The package interfaces with computational approaches to scientific discovery, including `scikit-learn` estimators for scientific model discovery, `sweetpea` for automated experimental design, `firebase_admin` for automated behavioral data collection, and `autodoc` for automated documentation of the empirical research process. While initially developed for the behavioral sciences, `autora` is designed as a general framework for closed-loop scientific discovery, with applications in other empirical sciences. Use cases of `autora` include the execution of closed-loop empirical studies [@musslick2024], the benchmarking of scientific discovery algorithms [@hewson_bayesian_2023], and the implementation of metascientific studies [@musslick_evaluation_2023]. 
+
+# Statement of Need
+The pace of empirical research is constrained by the rate at which scientists can alternate between the design and execution of experiments, on the one hand, and the derivation of scientific knowledge, on the other hand. However, attempts to increase this rate can compromise scientific rigor, leading to lower quality of formal modeling, insufficient documentation, and non-replicable findings. `autora` aims to surmount these limitations by formalizing the empirical research process and automating the generation, estimation, and empirical testing of scientific models. By providing a declarative language for empirical research, `autora` offers greater transparency and rigor in empirical research while accelerating scientific discovery. While existing scientific computing packages solve individual aspects of empirical research, there is no workflow mechanic for integrating them into a single pipeline, e.g., to enable closed-loop experiments. `autora` offers such a workflow mechanic, integrating Python packages for automating specific aspects of the empirical research process.
+
[…]
