+++ paper.bib
@@ -1,0 +1,104 @@
+@inproceedings{hewson_bayesian_2023,
+	title = {Bayesian machine scientist for model discovery in psychology},
+	author = {Hewson, Joshua and Strittmatter, Younes and Marinescu, Ioana and Williams, Chad and Musslick, Sebastian},
+	year = {2023},
+    booktitle = {NeurIPS 2023 AI for Science Workshop},
+    url={https://openreview.net/forum?id=XHFfvzlQ1n}
+}
+
+@inproceedings{musslick_evaluation_2023,
+	title = {An evaluation of experimental sampling strategies for autonomous empirical research in cognitive science},
+	volume = {45},
+	author = {Musslick, Sebastian and Hewson, Joshua TS and Andrew, Benjamin W and Strittmatter, Younes and Williams, Chad C and Dang, George T and Dubova, Marina and Holland, John Gerrard},
+	year = {2023},
+	note = {Issue: 45},
+}
+
+
+@article{dubova_against_2022,
+	title = {Against theory-motivated experimentation in science},
+	volume = {24},
+	journal = {MetaArXiv. June},
+	author = {Dubova, Marina and Moskvichev, Arseny and Zollman, Kevin},
+	year = {2022},
+	doi = {10.31222/osf.io/ysv2u},
+}
+
+@article{palan_prolific_2018,
+	title = {Prolific. ac—{A} subject pool for online experiments},
+	volume = {17},
+	issn = {2214-6350},
+	journal = {Journal of Behavioral and Experimental Finance},
+	author = {Palan, Stefan and Schitter, Christian},
+	year = {2018},
+	note = {Publisher: Elsevier},
+	pages = {22--27},
+	doi = {10.1016/j.jbef.2017.12.004},
+}
+
+@article{pedregosa2011scikit,
+  title={Scikit-learn: Machine learning in python},
+  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
+  journal={Journal of Machine Learning Research},
+  volume={12},
+  pages={2825--2830},
+  year={2011},
+  publisher={JMLR. org}
+}
+
+@inproceedings{
+petersen2021deep,
+title={Deep symbolic regression: Recovering mathematical expressions from data via risk-seeking policy gradients},
+author={Brenden K Petersen and Mikel Landajuela Larma and Terrell N. Mundhenk and Claudio Prata Santiago and Soo Kyung Kim and Joanne Taery Kim},
+booktitle={International Conference on Learning Representations},
+year={2021},
+url={https://openreview.net/forum?id=m5Qsh0kBQG},
+doi={10.48550/arXiv.1912.04871}
+}
+
+
+@inproceedings{landajuela_unified_2022,
+ author = {Landajuela, Mikel and Lee, Chak Shing and Yang, Jiachen and Glatt, Ruben and Santiago, Claudio P and Aravena, Ignacio and Mundhenk, Terrell and Mulcahy, Garrett and Petersen, Brenden K},
+ booktitle = {Advances in Neural Information Processing Systems},
+ editor = {S. Koyejo and S. Mohamed and A. Agarwal and D. Belgrave and K. Cho and A. Oh},
+ pages = {33985--33998},
+ publisher = {Curran Associates, Inc.},
+ title = {A Unified Framework for Deep Symbolic Regression},
[…]
