### 2024-05-28 9870399
extended README to include description of packages
