+++ README.md
@@ -30,5 +30,8 @@ Experimentalist Components:
 - [autora[experimentalist-uncertainty]](https://github.com/AutoResearch/autora-experimentalist-uncertainty): Automated experimental design based on model uncertainty.
 
 Experiment Runner Components:
+- [autora[experiment-runner-firebase-prolific]](https://github.com/AutoResearch/autora-experiment-runner-firebase-prolific): Automated data collection through serving an experiment on [Firebase](https://firebase.google.com/) and recruiting participants with [Prolific](https://www.prolific.com/).
+- [autora[experiment-runner-experimentation-manager-firebase]](https://github.com/AutoResearch/autora-experiment-runner-experimentation-manager-firebase): Automated data collection through serving an experiment on Firebase.
+- [autora[experiment-runner-recruitment-manager-prolific]](https://github.com/AutoResearch/autora-experiment-runner-recruitment-manager-prolific): Automated participant recruitment with Prolific.
 
 
