### 2024-05-29 98f6970
changed paper compilation to occur on push
