+++ README.md
@@ -7,8 +7,8 @@ This repository hosts the joint JOSS submission of the [AutoRA](https://github.c
 The AutoRA framework consists of core packages with minimal dependencies.
 
 - [autora](https://github.com/AutoResearch/autora): Main repository that hosts the documentation of autora across all vetted subpackages and serves as the main package.
-- [autora-core](https://github.com/AutoResearch/autora): AutoRA workflow mechanics.
-- [autora-synthetic](https://github.com/AutoResearch/autora): Synthetic models for benchmarking in AutoRA.
+- [autora-core](https://github.com/AutoResearch/autora-core): AutoRA workflow mechanics.
+- [autora-synthetic](https://github.com/AutoResearch/autora-synthetic): Synthetic models for benchmarking in AutoRA.
 
 **OPTIONAL PACKAGES** (optional dependencies):
 
@@ -17,13 +17,13 @@ In addition, the AutoRA framework provides vetted optional dependencies for work
 Theorist Components:
 - [autora[theorist-darts]](https://github.com/AutoResearch/autora-theorist-darts): Automated model discovery with Differentiable Architecture Search. 
 - [autora[theorist-bms]](https://github.com/AutoResearch/autora-theorist-bms): Automated model discovery with Bayesian Machine Scientist.
-- [autora[theorist-bsr]](https://github.com/AutoResearch/autora-theorist-sr): Automated model discovery with Bayesian Symbolic Regression.
+- [autora[theorist-bsr]](https://github.com/AutoResearch/autora-theorist-bsr): Automated model discovery with Bayesian Symbolic Regression.
 
 Experimentalist Components:
-- [autora[experimentalist-falsification]](https://github.com/AutoResearch/autora-experimentalist-uncertainty): Automated experimental design based on predicted model falsification.
+- [autora[experimentalist-falsification]](https://github.com/AutoResearch/autora-experimentalist-falsification): Automated experimental design based on predicted model falsification.
 - [autora[experimentalist-inequality]](https://github.com/AutoResearch/autora-experimentalist-inequality): Automated experimental design based on a pairwise distance metric.
-- [autora[experimentalist-leverage]](https://github.com/AutoResearch/autora-experimentalist-uncertainty): Automated experimental design based on the leverage of data points on model predictions.
-- [autora[experimentalist-mixture]](https://github.com/AutoResearch/autora-experimentalist-uncertainty): Automated experimental design based on mixture of  experimentalists.
+- [autora[experimentalist-leverage]](https://github.com/AutoResearch/autora-experimentalist-leverage): Automated experimental design based on the leverage of data points on model predictions.
+- [autora[experimentalist-mixture]](https://github.com/AutoResearch/autora-experimentalist-mixture): Automated experimental design based on mixture of  experimentalists.
 - [autora[experimentalist-nearest-value]](https://github.com/AutoResearch/autora-experimentalist-nearest-value): Automated experimental design based on distance to existing pool of experiment conditions.
 - [autora[experimentalist-novelty]](https://github.com/AutoResearch/autora-experimentalist-novelty): Automated experimental design based on experiment novelty.
 - [autora[experimentalist-model-disagreement]](https://github.com/AutoResearch/autora-experimentalist-model-disagreement): Automated experimental design based on model disagreement.
