### 2024-05-29 #6769 @musslick
@jbytecode

Thank you for pointing us in the right direction.

We opened a third repository for the submission only: [https://github.com/AutoResearch/autora-paper](https://github.com/AutoResearch/autora-paper)

In this repository, we link the two main packages `autora-core` and `autora-synthetic`, as well as the main package `autora` which provides a common namespace for installation of and documentation for all sub packages.

The AutoRA framework includes a number of optional dependencies that we also listed in the linked submission, primarily to illustrate the scope of the code contribution. However, we agree that reviewing `autora-core` and `autora-synthetic` would be most appropriate (potentially together with the `autora` repository).

We believe that our submission falls within the expertise of the following reviewers:
- Sean Devine ([seandamiandevine](https://github.com/seandamiandevine))
- Benedikt Ehinger ([behinger](https://github.com/behinger))

Thanks again for your help!
