### 2024-05-29 #6769 @jbytecode
👋👋👋 Dear @seandamiandevine, @behinger 👋👋👋

[[AutoRA: Automated Research Assistant for Closed-Loop Empirical Research](https://github.com/openjournals/joss-reviews/issues/6769#top)

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6769).

Thank you in advance!
