### 2024-06-04 #6839 @editorialbot
[REVIEW]: AutoRA: Automated Research Assistant for Closed-Loop Empirical Research
**Submitting author:** @musslick (<a href="http://orcid.org/0000-0002-8896-639X">Sebastian Musslick</a>)
**Repository:** https://github.com/AutoResearch/autora-paper
**Branch with paper.md** (empty if default branch): main
**Version:** v4.2.0
