### 2024-06-07 b5da18f
updated acknowledgement for auxiliary packages which are currently under development (e.g. autodoc).
