### 2024-06-21 #6839 @musslick
@jbytecode Sure thing!

@szorowi1 (also tagging @seandamiandevine ) Thanks for checking about this. We discussed with the development team what might be a good functionality test for AutoRA. We reached consensus that the [Tutorials](https://autoresearch.github.io/autora/tutorials/) and the two examples [Equation Discovery](https://autoresearch.github.io/autora/theorist/Equation%20Discovery/) and [Online Closed-Loop Discovery](https://autoresearch.github.io/autora/user-cookiecutter/docs/) would capture most of the core functionality of AutoRA. So evaluating those might be most appropriate for a functionality test. Note that all of the tutorials (except the online closed-loop discovery) should be executable via Google Colab. Please let us know if you run into any issues or have any other questions---and thanks for all your work!
