### 2024-07-03 #6839 @szorowi1
Apologies to all for the delay, it's been a hectic few weeks!

Let me start by saying congrats to @musslick and co-authors/collaborators! This is a really impressive framework and it's obvious how much careful attention, thought, and effort went into developing it. Kudos!

I've now had a chance to work through the documentation, tutorials, and examples. The installation went fine, the code works as expected, and the Docs/API are robust. To echo @seandamiandevine, I also ran into a number of errors when running through the `Equation Discovery` tutorial in Collab having to do with data shape mismatches. When running the `Experimentalist.ipynb` tutorial notebook in Collab, I also ran into the following error early on:

```
[<ipython-input-2-508cdcdb2e51>](https://localhost:8080/#) in <cell line: 4>()
      2 from sklearn.linear_model import LinearRegression
      3 from autora.variable import DV, IV, ValueType, VariableCollection
----> 4 from autora.experimentalist.sampler.falsification import falsification_sample, falsification_score_sample

ModuleNotFoundError: No module named 'autora.experimentalist.sampler'
```
I agree with @seandamiandevine it would be good to make sure the Example notebooks run all the way through for new users.

Some general thoughts, none of which should necessarily preclude acceptance or publication:

- Depending on who the intended "core audience" is, I think more elaborate examples might be helpful. If one of the main intended audiences are cognitive scientists/psychologists, then an additional example might be using AutoRA to identify experiments best suited for discriminating between competing cognitive models (e.g., as in [Cavagnaro et al. 2013](https://link.springer.com/article/10.1007/s11166-013-9179-3) or [Cavagnaro et al. 2016](https://link.springer.com/article/10.1007/s11166-016-9242-y)). [As an aside, it would be interesting to hear how you think AutoRA relates to adaptive design optimization more broadly.]
- Relatedly, one experimentalist type not yet covered (I don't think?) is choosing trials to minimize posterior uncertainty over model parameters (e.g., [Ahn et al. 2020](https://doi.org/10.1038/s41598-020-68587-x)).
- Final thought: Oh what I would have given to have had something like this during my PhD 🙂
