### 2024-07-05 #6839 @musslick
Dear @seandamiandevine and @szorowi1,

Thank you both so much for investing the time and effort into this review and for providing such thorough and constructive feedback. We really appreciate that!

I discussed your feedback with the team, and we agree that there is not sufficient (and complete) information about how to utilize the closed-loop functionality of AutoRA for *real-world* experiments. Adding respective examples in the documentation would be beneficial, especially for researchers interested in behavioral experiments.

We propose to do the following:
1. **Fix the errors** in the `Equation Discovery` tutorial and the `Experimentalist.ipynb` notebook.
2. **Include the following two end-to-end examples for closed-loop experimentation** with AutoRA (using a both _Prolific_, and _Firebase_):

	2.1  _Mathematical model discovery_ for a psychophysics experiment

	2.2  _Computational (reinforcement learning) model discovery_ for a one-armed bandit experiment

    Once we implemented and internally vetted those tutorials we would love to get your feedback on those. That said, we would also understand if you've had enough of AutoRA already and/or don't have the time ;)

    As a quick fix, we already expanded the [Closed-Loop Online Experiment Example](https://autoresearch.github.io/autora/user-cookiecutter/docs/) to include a description of how to combine AutoRA with Prolific (to address @seandamiandevine' initial point).

    In addition, to follow-up on the general thoughts from @szorowi1's, we would aim to include *two additional examples for closed-loop experimentation* (also using Prolific and Firebase). We may not be able to get them implemented over the course of the review process but wanted to hear your thoughts on whether these could be a useful target for our next development milestone:

	2.3  _Drift diffusion model comparison_ for a random-object-kinematogram (RDK) experiment using Bayesian optimal experimental design (specifically minimizing posterior uncertainty)

	2.4 _Experiment parameter tuning_ for a task-switching experiment (to illustrate how AutoRA can be used to for automated design optimization, e.g., to enhance a desired behavioral effect, such as task switch costs)

Finally, to address @szorowi1's question: We think that AutoRA could be used for design optimization (we could illustrate this in Example 2.4). However, it's not (yet) capable of adapting the experiment on the fly, i.e., within a single experiment session. Rather, it can help optimize design after collecting data from a set of experiments, and then proposing a new set of experiments.

Please let us know what you think!
