### 2024-10-02 #6839 @musslick
> @musslick - May I request an update please? Sorry if I am bothering. Thank you in advance.

Thanks for checking in. We are aiming to have the new release up by end of next week, and will ping you!
