### 2024-10-20 #6839 @musslick
Dear @jbytecode @szorowi1  @seandamiandevine

Thank you so much for your patience with this revision. Your feedback has been incredibly valuable, helping us uncover multiple inconsistencies in the AutoRA documentation and concomitant opportunities for improvement. This led us to address some deeper issues in the core code base, which is why the revision process took a bit longer than anticipated.

Following your suggestions, we ended up re-structuring the tutorials into _Basic Tutorials_ and _Use Case Tutorials_. The most relevant changes include:
- Fixing the [Theorist Tutorial](https://autoresearch.github.io/autora/tutorials/basic/Tutorial%20Ia%20Theorists/), which can now be found in the Basic Tutorials section.
- Fixing the [Experimentalist Tutorial](https://autoresearch.github.io/autora/tutorials/basic/Tutorial%20Ib%20Experimentalists/), which can now also be found in the Basic Tutorials section.
- Adding an end-to-end example for a [Closed-Loop Psychophysics Study](https://autoresearch.github.io/autora/examples/closed-loop-basic/).
- Adding an end-to-end example for a [Closed-Loop One-Armed Bandit Study](https://autoresearch.github.io/autora/examples/closed-loop-bandit-task/).

We invite you to have a look at these revised sections.

*To ease your review:* Running the revised tutorials shouldn't require any local installation. You should be able to execute both Basic Tutorials in Google Collab. In addition, you should be able run both Use Case Tutorials within GitHub Codespaces (we provide guidelines in the respective tutorials).

Thank you again for your your valuable time and effort. We look forward to your feedback!
