### 2024-11-24 b80b0c5
added automation perspective as a ref
