+++ paper.bib
@@ -69,6 +69,17 @@ doi={10.48550/arXiv.1912.04871}
  year = {2022}
 }
 
+@article{musslick2024perspective,
+  author = {Musslick, Sebastian and Bartlett, Laura K. and Chandramouli, Suyog H. and Dubova, Marina and Gobet, Fernand and Griffiths, Thomas L. and Hullman, Jessica and King, Ross D. and Kutz, J. Nathan and Lucas, Christopher G. and Mahesh, Suhas and Pestilli, Franco and Sloman, Sabina J. and Holmes, William R.},
+  title = {Automating the Practice of Science--Opportunities, Challenges, and Implications},
+  journal = {Proceedings of the National Academy of Sciences},
+  year = {2024},
+  note = {In press},
+  eprint = {2409.05890},
+  archivePrefix = {arXiv},
+  primaryClass = {cs.AI},
+}
+
 
 @article{cranmer_discovering_2020,
 	title = {Discovering symbolic models from deep learning with inductive biases},
