### 2024-11-24 0e1c65d
added Weinhardt et al as a ref
