+++ paper.bib
@@ -56,6 +56,14 @@ url={https://openreview.net/forum?id=m5Qsh0kBQG},
 doi={10.48550/arXiv.1912.04871}
 }
 
+@inproceedings{
+weinhardt2024computational,
+title={Computational discovery of human reinforcement learning dynamics from choice behavior},
+author={Daniel Weinhardt and Maria K Eckstein and Sebastian Musslick},
+booktitle={NeurIPS 2024 Workshop on Behavioral Machine Learning},
+year={2024},
+url={https://openreview.net/forum?id=x2WDZrpgmB}
+}
 
 @inproceedings{landajuela_unified_2022,
  author = {Landajuela, Mikel and Lee, Chak Shing and Yang, Jiachen and Glatt, Ruben and Santiago, Claudio P and Aravena, Ignacio and Mundhenk, Terrell and Mulcahy, Garrett and Petersen, Brenden K},
