### 2024-11-24 6b443a6
discussed LLMs as theorists
