+++ paper.bib
@@ -56,6 +56,13 @@ url={https://openreview.net/forum?id=m5Qsh0kBQG},
 doi={10.48550/arXiv.1912.04871}
 }
 
+@article{binz2024centaur,
+  title={Centaur: a foundation model of human cognition},
+  author={Binz, Marcel and Akata, Elif and Bethge, Matthias and Br{\"a}ndle, Franziska and Callaway, Fred and Coda-Forno, Julian and Dayan, Peter and Demircan, Can and Eckstein, Maria K and {\'E}ltet{\H{o}}, No{\'e}mi and others},
+  journal={arXiv preprint arXiv:2410.20268},
+  year={2024}
+}
+
 @inproceedings{
 weinhardt2024computational,
 title={Computational discovery of human reinforcement learning dynamics from choice behavior},
