### 2024-11-24 5e4cb15
clarified role of synthetic experiment runners as simulators
