+++ paper.bib
@@ -88,8 +88,7 @@ url={https://openreview.net/forum?id=x2WDZrpgmB}
   author = {Musslick, Sebastian and Bartlett, Laura K. and Chandramouli, Suyog H. and Dubova, Marina and Gobet, Fernand and Griffiths, Thomas L. and Hullman, Jessica and King, Ross D. and Kutz, J. Nathan and Lucas, Christopher G. and Mahesh, Suhas and Pestilli, Franco and Sloman, Sabina J. and Holmes, William R.},
   title = {Automating the Practice of Science--Opportunities, Challenges, and Implications},
   journal = {Proceedings of the National Academy of Sciences},
-  year = {2024},
-  note = {In press},
+  year = {in press},
   eprint = {2409.05890},
   archivePrefix = {arXiv},
   primaryClass = {cs.AI},
