+++ paper.bib
@@ -18,7 +18,7 @@
 @article{dubova_against_2022,
 	title = {Against theory-motivated experimentation in science},
 	volume = {24},
-	journal = {MetaArXiv. June},
+	journal = {MetaArXiv},
 	author = {Dubova, Marina and Moskvichev, Arseny and Zollman, Kevin},
 	year = {2022},
 	doi = {10.31222/osf.io/ysv2u},
@@ -60,7 +60,8 @@ doi={10.48550/arXiv.1912.04871}
   title={Centaur: a foundation model of human cognition},
   author={Binz, Marcel and Akata, Elif and Bethge, Matthias and Br{\"a}ndle, Franziska and Callaway, Fred and Coda-Forno, Julian and Dayan, Peter and Demircan, Can and Eckstein, Maria K and {\'E}ltet{\H{o}}, No{\'e}mi and others},
   journal={arXiv preprint arXiv:2410.20268},
-  year={2024}
+  year={2024},
+  doi={10.48550/arXiv.2410.20268}
 }
 
 @inproceedings{
