### 2024-11-26 #6839 @seandamiandevine
@jbytecode @musslick

Thank you for the revisions! After rerunning the tutorial, everything runs well for me. I also double-checked the local install and that also works for me.

The new Use Case examples are excellent and clearly address my initial concerns. They also serve as a good introduction to Firebase for researchers who are new to server-side development (if it can be called that).

Overall, I find that my comments have been addressed and I'm happy to recommend publication in the current form. @jbytecode, please let me know if I need to do anything else at this stage.

Thanks again for thinking of me as a reviewer and congratulations to @musslick and the team for their contribution!
