### 2024-12-01 #6839 @szorowi1
Hi all, thanks for your patience. I've had a chance now to revise the revised introductory tutorials and the 2-arm bandit use case tutorial. The former look fantastic and run as expected. The latter is simply awesome. I'm very excited to see the connection to the tiny-RNNs work for closed-loop modeling of human RL behavior.

I'm with @seandamiandevine. All of my comments have been addressed and I recommend acceptance of this manuscript. Congrats, @musslick!
