### 2024-12-03 #6839 @jbytecode
@musslick - I've took a look at the manuscript, indeed, it seems okay to me.

We need to create a tagged release of the repository and create a Zenodo archive using this release.

Please use the correct order of authors with their DOIs in the Zenodo metadata and all of the metadata in the manuscript should match those in the archive.

After finishing the stuff (as formerly indicated and listed in the post above) please report me the tagged release, e.g. (in form of v4.0.0), and the Zenodo archive.

Thank you in advance.
