### 2024-12-04 #6839 @musslick
@szorowi1 Thank you so much for taking the time and effort to re-review the documentation! We are excited to hear that you liked the closed-loop bandits example. Indeed, we I have been chatting with Marcello about tiny-RNNs and we are considering implementing those as theorists in AutoRA.
Overall, your and @seandamiandevine's review have been incredibly constructive and valuable for us! Thanks again!

@jbytecode That sounds great. I will work on the checklist and ping you once it is finished.

@musslick
- [ ] Fix missing DOIs in the paper
