### 2024-12-04 #6839 @musslick
@jbytecode Alright, it's all done.

This is the new version number: **v4.2.0**

And here is the corresponding doi from Zenodo: [10.5281/zenodo.14278950](https://doi.org/10.5281/zenodo.14278950)

(I assume you prefer the doi associated with the release, as listed above. Zenodo also has a doi for all (future) versions of the repo if that is preferred: [10.5281/zenodo.10277414](https://doi.org/10.5281/zenodo.10277414)).
