### 2024-12-05 #6839 @jbytecode
@musslick - The DOI 10.5555/1953048.2078195 doesn't resolve, could you please check it?
