+++ paper.bib
@@ -43,7 +43,6 @@
   volume={12},
   pages={2825--2830},
   year={2011},
-  doi={10.5555/1953048.2078195},
   publisher={JMLR. org}
 }
 
