### 2024-12-05 #6839 @musslick
@jbytecode Thank you for the fast response! I removed the DOI (paper should be updated now).

Even though it is listed in [some places](https://dblp.org/rec/journals/jmlr/PedregosaVGMTGBPWDVPCBPD11.html) for the reference in question, it [doesn't resolve](https://doi.org/10.5555/1953048.2078195) through the DOI database. So it might be best to leave it out.
