### 2024-12-05 #6839 @jbytecode
@musslick - I've just sent a pull request to fix a small typo in the bibliography. Please review the PR. If it's okay to you, please merge. (https://github.com/AutoResearch/autora-paper/pull/3)

The paper and the bibliography look good to me. After reviewing the PR, please perform a full read of the manuscript. Please ping me when you've done with it.

Thank you in advance.
