+++ paper.bib
@@ -11,6 +11,7 @@
 	volume = {45},
 	author = {Musslick, Sebastian and Hewson, Joshua TS and Andrew, Benjamin W and Strittmatter, Younes and Williams, Chad C and Dang, George T and Dubova, Marina and Holland, John Gerrard},
 	year = {2023},
+	pages = {1386--1392},
 	note = {Issue: 45},
 }
 
