### 2024-12-05 6d4aab6
improved figure caption
