### 2024-12-05 c82466b
fixed minor wording and acknowledgements
