+++ paper.bib
@@ -91,9 +91,6 @@ url={https://openreview.net/forum?id=x2WDZrpgmB}
   title = {Automating the Practice of Science--Opportunities, Challenges, and Implications},
   journal = {Proceedings of the National Academy of Sciences},
   year = {in press},
-  eprint = {2409.05890},
-  archivePrefix = {arXiv},
-  primaryClass = {cs.AI},
 }
 
 
