### 2024-12-05 #6839 @musslick
@jbytecode We performed a full read of the manuscript,  made some final edits and merged them into the main branch.
