### 2024-12-05 #6839 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
