### 2024-12-05 #6839 @musslick
@jbytecode @crvernon Thank you for this incredibly constructive and responsive editing & publication process. This was our first time submitting to JOSS and we are thrilled about the journal!
